+++ NEWS.md
@@ -10,6 +10,12 @@
   `gglikert_stacked()` (using `hex_bw()`) (#57)
 * new argument `data_fun` for `gglikert()`, `gglikert_data()` and
   `gglikert_stacked()` (#60)
+* new scale `scale_fill_likert()`
+* new argument `cutoff` for `gglikert()`, `position_likert()` and
+  `scale_fill_likert()` (#64)
+* new helper `pal_extender()` and corresponding `scale_fill_extender()` and
+  `scale_colour_extender()`
+* new sorting option `"prop_lower"` for `gglikert()` (#62)
 
 # ggstats 0.5.1